Pull TrinoException subclasses to top level classes
Move TrinoException subclasses not to be nested classes. When a class is
nested, it's human readable name is `package.Enclosing.MyClass` but the
actual class name is `package.Enclosing$MyClass`. Thus, when a
`TrinoException` is a nested class, the
`io.trino.client.FailureInfo.type` cannot used to restore the original
exception class, which by needed in subsequent changes.
