+++ core/trino-main/src/main/java/io/trino/json/PathEvaluationVisitor.java
@@ -51,8 +51,8 @@ import io.trino.json.ir.IrPathNode;
 import io.trino.json.ir.IrPredicateCurrentItemVariable;
 import io.trino.json.ir.IrSizeMethod;
 import io.trino.json.ir.IrTypeMethod;
+import io.trino.json.ir.JsonLiteralConversionException;
 import io.trino.json.ir.SqlJsonLiteralConverter;
-import io.trino.json.ir.SqlJsonLiteralConverter.JsonLiteralConversionException;
 import io.trino.json.ir.TypedValue;
 import io.trino.spi.function.OperatorType;
 import io.trino.spi.type.CharType;
