+++ core/trino-main/src/test/java/io/trino/json/ir/TestSqlJsonLiteralConverter.java
@@ -29,7 +29,6 @@ import com.fasterxml.jackson.databind.node.NullNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.databind.node.ShortNode;
 import com.fasterxml.jackson.databind.node.TextNode;
-import io.trino.json.ir.SqlJsonLiteralConverter.JsonLiteralConversionException;
 import io.trino.spi.type.Int128;
 import org.assertj.core.api.AssertProvider;
 import org.assertj.core.api.RecursiveComparisonAssert;
